further refactored main dashboard by making last updated session title its own component
