+++ jada-app/src/Components/LastUpdatedSession/lastUpdatedSession.css
@@ -1,0 +1,3 @@
+.lastSessionRun {
+    margin: 40px 0;
+}
