--- jada-app/src/Components/MainDashboard/mainDashboard.css
@@ -1,7 +1,0 @@
-.lastSessionRun {
-    margin: 40px 0;
-}
-
-.lastSessionRun span {
-    font-weight: 300;
-}
